requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pydantic==2.5.0
sqlite3-wrapper==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
redis==5.0.1
asyncio-mqtt==0.16.1
schedule==1.2.0
psutil==5.9.6
python-dotenv==1.0.0